sso/requirements.txt
blinker==1.8.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-auth==1.27.0
google-auth-oauthlib==0.4.2
greenlet==3.0.3
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==1.26.3
Werkzeug==3.0.3
